+++ lib/td_dq/rules/rule_results/bulk_load.ex
@@ -7,6 +7,7 @@ defmodule TdDq.Rules.RuleResults.BulkLoad do
 
   alias Ecto.Multi
   alias TdDd.Repo
+  alias TdDq.Cache.ImplementationLoader
   alias TdDq.Cache.RuleLoader
   alias TdDq.Implementations.Implementation
   alias TdDq.Rules.Audit
@@ -28,6 +29,12 @@ defmodule TdDq.Rules.RuleResults.BulkLoad do
     Multi.new()
     |> Multi.run(:ids, fn _, _ -> bulk_insert(records) end)
     |> Multi.run(:results, fn _, %{ids: ids} -> RuleResults.select_results(ids) end)
+    |> Multi.run(:cache, fn _, %{results: results} ->
+      {:ok,
+       Enum.map(results, fn %{implementation_id: implementation_id} ->
+         ImplementationLoader.maybe_update_implementation_cache(implementation_id)
+       end)}
+    end)
     |> Multi.run(:audit, Audit, :rule_results_created, [0])
     |> Repo.transaction()
     |> bulk_refresh()