+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-4271] Support for linking implementations with business concepts
+
 ## [4.40.0] 2022-03-14
 
 ### Changed
@@ -9,6 +15,7 @@
 - [TD-4491] Refactored search and permissions
 - [TD-4604] Include actions in `POST /api/rule_implementations/search` and `POST
  /api/data_structures/search` responses
+
 
 ## [4.39.0] 2022-03-07